% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deepdep.R
\name{add_layers_to_vertices}
\alias{add_layers_to_vertices}
\title{Add layer property to graph vertices}
\usage{
add_layers_to_vertices(G)
}
\arguments{
\item{G}{An \code{igraph} object}
}
\description{
Add layer property to graph vertices
}
