% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepdep.R
\name{deepdep}
\alias{deepdep}
\title{Visualizing dependiences between packages}
\usage{
deepdep(package, depth = 3)
}
\arguments{
\item{package}{\code{character} Name of the package that is on CRAN}

\item{depth}{\code{integer}}
}
\description{
Visualize connections between packages.
}
