% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deepdep.R
\name{delete_edges_within_layer}
\alias{delete_edges_within_layer}
\title{Remove edges on the same level}
\usage{
delete_edges_within_layer(G)
}
\arguments{
\item{G}{An \code{igraph} object}
}
\description{
Remove edges on the same level
}
