% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependencies.R
\name{get_dependencies}
\alias{get_dependencies}
\alias{print.package_dependencies}
\title{Acquire dependencies of the package}
\usage{
get_dependencies(package, downloads = FALSE)

\method{print}{package_dependencies}(x, ...)
}
\arguments{
\item{package}{A \code{character}. Name of the package that is on CRAN.}

\item{downloads}{A \code{logical}. If \code{TRUE} add package downloads data. By default it's \code{FALSE}.}
}
\value{
An object of \code{package_dependencies} class.
}
\description{
This function is using \code{\link{get_description}} and \code{\link{get_downloads}}
to acquire dependencies.
}
\examples{
library(deepdep)

dependencies <- get_dependencies("ggplot2")
dependencies


}
\author{
Hubert Baniecki, Szymon Maksymiuk
}
