% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_description.R
\name{get_description}
\alias{get_description}
\alias{print.package_description}
\title{Scrap the DESCRIPTION file and CRAN metadata of the package}
\usage{
get_description(package)

\method{print}{package_description}(x, ...)
}
\arguments{
\item{package}{A \code{character}. Name of the package that is on CRAN.}
}
\value{
An object of \code{package_description} class.
}
\description{
[API](https://github.com/r-hub/crandb) [CRAN Data Base](http://crandb.r-pkg.org)
}
\examples{
library(deepdep)

desc <- get_description("ggplot2")
print(desc)


}
\author{
Hubert Baniecki, Szymon Maksymiuk
}
