% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_downloads.R
\name{get_downloads}
\alias{get_downloads}
\title{Scrap the download data of the package}
\usage{
get_downloads(package)
}
\arguments{
\item{package}{A \code{character}. Name of the package that is on CRAN.}
}
\value{
An object of \code{package_downloads} class.
}
\description{
This function uses [API](https://github.com/r-hub/cranlogs) of [CRAN Logs](http://cran-logs.rstudio.com/)
to scrap the download logs of the package.
}
\examples{
library(deepdep)

downloads <- get_downloads("ggplot2")
downloads


}
