% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_utils.R
\name{gravity_child}
\alias{gravity_child}
\title{Push parent towards the child}
\usage{
gravity_child(node, child_node)
}
\arguments{
\item{node}{A row from \code{layout} data frame with node to push towards
the child.}

\item{child_node}{A row from \code{layout} data frame with child node.}
}
\description{
Push parent towards the child
}
