% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_utils.R
\name{gravity_iteration}
\alias{gravity_iteration}
\title{Run gravity iteration}
\usage{
gravity_iteration(layout, G)
}
\arguments{
\item{layout}{Data frame object containing data about current layout.}

\item{G}{Graph which layout is being adjusted.}
}
\description{
Run gravity iteration
}
