% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_utils.R
\name{gravity_node}
\alias{gravity_node}
\title{Compute gravity on a node}
\usage{
gravity_node(node, layout, G)
}
\arguments{
\item{node}{A row from \code{layout} data frame with node to gravitate.}

\item{layout}{Data frame object containing data about current layout.}

\item{G}{Graph which layout is being adjusted.}
}
\description{
Compute gravity on a node
}
