% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_utils.R
\name{gravity_parent}
\alias{gravity_parent}
\title{Push child towards the parent}
\usage{
gravity_parent(node, parent_node)
}
\arguments{
\item{node}{A row from \code{layout} data frame with node to push towards
the parent.}

\item{parent_node}{A row from \code{layout} data frame with parent node.}
}
\description{
Push child towards the parent
}
