% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_utils.R
\name{gravity_sibling}
\alias{gravity_sibling}
\title{Push parent from the sibling}
\usage{
gravity_sibling(node, sibling_node)
}
\arguments{
\item{node}{A row from \code{layout} data frame with node to push from the
sibling.}

\item{sibling_node}{A row from \code{layout} data frame with sibling node.}
}
\description{
Push parent from the sibling
}
