% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deepdep.R
\name{plot.deepdep}
\alias{plot.deepdep}
\title{plot}
\usage{
\method{plot}{deepdep}(x, plot_type, ...)
}
\arguments{
\item{x}{x}

\item{plot_type}{A \code{character}. \code{circular}, \code{ee}, \code{bb}
"horizontal tree" or "vertical tree".}

\item{...}{other}
}
\description{
plot
}
