% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepdep.R
\name{print.deepdep}
\alias{print.deepdep}
\title{Print function for an object of \code{deepdep} class.}
\usage{
\method{print}{deepdep}(x, ...)
}
\arguments{
\item{x}{An object of \code{deepdep} class.}

\item{...}{other}
}
\description{
Print function for an object of \code{deepdep} class.
}
\examples{
library(deepdep)

dd <- deepdep("stringr")
dd

}
