% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependencies.R
\name{print.package_dependencies}
\alias{print.package_dependencies}
\title{Print function for an object of \code{package_dependencies} class}
\usage{
\method{print}{package_dependencies}(x, ...)
}
\arguments{
\item{x}{An object of \code{package_dependencies} class.}

\item{...}{other}
}
\description{
Print function for an object of \code{package_dependencies} class
}
\examples{
library(deepdep)

dependencies <- get_dependencies("stringr")
dependencies

}
\author{
Hubert Baniecki, Szymon Maksymiuk
}
