% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_description.R
\name{print.package_description}
\alias{print.package_description}
\title{Print function for an object of \code{package_description} class}
\usage{
\method{print}{package_description}(x, ...)
}
\arguments{
\item{x}{An object of \code{package_description} class.}

\item{...}{other}
}
\description{
Print function for an object of \code{package_description} class
}
\examples{
library(deepdep)

desc <- get_description("stringr")
desc

}
\author{
Hubert Baniecki, Szymon Maksymiuk
}
