% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_downloads.R
\name{print.package_downloads}
\alias{print.package_downloads}
\title{Print function for an object of \code{package_downloads} class}
\usage{
\method{print}{package_downloads}(x, ...)
}
\arguments{
\item{x}{An object of \code{package_downloads} class.}

\item{...}{other}
}
\description{
Print function for an object of \code{package_downloads} class
}
\examples{
library(deepdep)

desc <- get_downloads("stringr")
desc

}
