% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_example.R
\name{run_shiny_example}
\alias{run_shiny_example}
\title{Run Shiny example for deepdep package}
\usage{
run_shiny_example()
}
\description{
This function runs Shiny app with example usage
of deepdep package.
}
