% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepdep.R
\name{deepdep}
\alias{deepdep}
\title{Visualizing dependiences between packages}
\usage{
deepdep(package, plot_type = "circular", deps_depth = "standard")
}
\arguments{
\item{package}{\code{\link{Character}} vector giving a package name.}

\item{plot_type}{\code{\link{Character}} string giving the type of the plot - either "circular",
"horizontal tree" or "vertical tree".}

\item{deps_depth}{\code{\link{Character}} string or \code{\link{numeric}} vector, giving 
information what types of dependencies to plot and how deep - see details.}
}
\description{
Visualize connections between packages.
}
