% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepdep.R
\name{deepdep}
\alias{deepdep}
\title{deepdep}
\usage{
deepdep(package, downloads = FALSE, depth = 1, bioc = FALSE,
  local = FALSE, deps_types = c("Depends", "Imports"))
}
\arguments{
\item{package}{A \code{character}. Name of the package that is on CRAN.}

\item{downloads}{A \code{logical}. If \code{TRUE} add package downloads data. By default it's \code{FALSE}.}

\item{depth}{An \code{integer}. Depth of package dependency. By default it's \code{1}.}

\item{bioc}{A \code{logical} value. Should Bioconductor dependencies descriptions be red from 
Bioconductor repository? For this option to work properly, \code{BiocManager} package needs to be 
installed.}

\item{deps_types}{A \code{character} vector. Types of dependencies that should be sought.
Possibilities are: Imports, Depends, Suggests, Enhances, LinkingTo. Defaults to 
\code(c\("Depends", "Imports"\))}
}
\value{
An object of \code{deepdep} class.
}
\description{
deepdep.
}
\examples{
library(deepdep)

dd <- deepdep("stringr", downloads = TRUE)
dd

dd2 <- deepdep("stringr", depth = 2)
dd2

}
\author{
Hubert Baniecki, Szymon Maksymiuk
}
