% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepdep.R
\name{deepdep}
\alias{deepdep}
\title{deepdep}
\usage{
deepdep(package, downloads = FALSE, depth = 1)
}
\arguments{
\item{package}{A \code{character}. Name of the package that is on CRAN.}

\item{downloads}{A \code{logical}. If \code{TRUE} add package downloads data. By default it's \code{FALSE}.}

\item{depth}{An \code{integer}. Depth of package dependency. By default it's \code{1}.}
}
\value{
An object of \code{deepdep} class.
}
\description{
deepdep.
}
\examples{
library(deepdep)

dd <- deepdep("stringr", downloads = TRUE)
dd

dd2 <- deepdep("stringr", depth = 2)
dd2

}
\author{
Hubert Baniecki, Szymon Maksymiuk
}
