% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependencies.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Acquire dependencies of the package}
\usage{
get_dependencies(package, downloads = FALSE, bioc = FALSE,
  local = FALSE, deps_types = c("Depends", "Imports"))
}
\arguments{
\item{package}{A \code{character}. Name of the package that is on CRAN.}

\item{downloads}{A \code{logical}. If \code{TRUE} add package downloads data. By default it's \code{FALSE}.}

\item{bioc}{A \code{logical} value. Should Bioconductor dependencies descriptions be red from 
Bioconductor repository? For this option to work properly, \code{BiocManager} package needs to be 
installed.}

\item{local}{A \code{logical} value. Should only already installed packages be checked?}

\item{deps_types}{A \code{character} vector. Types of dependencies that should be sought.
Possibilities are: Imports, Depends, Suggests, Enhances, LinkingTo. Defaults to 
\code(c\("Depends", "Imports"\))}
}
\value{
An object of \code{package_dependencies} class.
}
\description{
This function is using \code{\link{get_description}} and \code{\link{get_downloads}}
to acquire dependencies.
}
\examples{
library(deepdep)

dependencies <- get_dependencies("stringr")
dependencies


}
\author{
Hubert Baniecki, Szymon Maksymiuk
}
