% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_description.R
\name{get_description}
\alias{get_description}
\title{Scrap the DESCRIPTION file and CRAN metadata of the package}
\usage{
get_description(package, bioc = FALSE, local = FALSE,
  reset_cache = FALSE)
}
\arguments{
\item{package}{A \code{character}. Name of the package that is on CRAN.}

\item{bioc}{A \code{logical} value. Should Bioconductor dependencies descriptions be red from
Bioconductor repository? For this option to work properly, \code{BiocManager} package needs to be
installed.}

\item{local}{A \code{logical} value. Should only already installed packages be checked?}

\item{reset_cache}{A \code{logical} value. Should cache be cleared before obtaining the
list of packages?}
}
\value{
An object of \code{package_description} class.
}
\description{
[API](https://github.com/r-hub/crandb) [CRAN Data Base](http://crandb.r-pkg.org)
}
\examples{
library(deepdep)

desc <- get_description("stringr")
desc


}
