% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_downloads.R
\name{get_downloads}
\alias{get_downloads}
\alias{print.package_downloads}
\title{Scrap the download data of the package}
\usage{
get_downloads(package)

\method{print}{package_downloads}(x, ...)
}
\arguments{
\item{package}{A \code{character}. Name of the package that is on CRAN}
}
\value{
An object of \code{package_downloads} class
}
\description{
Download logs from the RStudio CRAN mirror through [META CRAN Logs](https://cranlogs.r-pkg.org/)
}
\examples{
library(deepdep)

downloads <- get_downloads("ggplot2")
downloads


}
\author{
Hubert Baniecki, Szymon Maksymiuk
}
