% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deepdep.R
\name{plot_dependencies}
\alias{plot_dependencies}
\alias{plot_dependencies.default}
\alias{plot_dependencies.deepdep}
\title{Plot dependency data}
\usage{
plot_dependencies(x, plot_type = "circular", same_level = FALSE)

\method{plot_dependencies}{default}(x, plot_type = "circular",
  same_level = FALSE)

\method{plot_dependencies}{deepdep}(x, plot_type = "circular",
  same_level = FALSE, n_iter = 10, ...)
}
\arguments{
\item{x}{A \code{\link{deepdep}} object.}

\item{plot_type}{A \code{character}. Possible values are \code{circular} and \code{tree}.}

\item{same_level}{A \code{boolean}. Whether to plot links between dependencies on the same
level.}

\item{n_iter}{Number of iterations made to optimize layout.}

\item{...}{Other arguments passed to plotting function.}
}
\description{
Visualize dependency data passed as \code{deepdep} object using
\code{ggplot2} and \code{ggraph} packages. Several tree-like layouts are
available
}
\examples{
library(deepdep)

dd <- deepdep("stringr")
plot_dependencies(dd, "tree")

dd2 <- deepdep("cranlogs", depth = 2)
plot_dependencies(dd2, "circular")
}
\author{
Mateusz Bąkała
}
