% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deepdep.R
\name{plot_dependencies}
\alias{plot_dependencies}
\alias{plot_dependencies.default}
\alias{plot_dependencies.character}
\alias{plot_dependencies.deepdep}
\title{Main plot function for a \code{deepdep} object.}
\usage{
plot_dependencies(x, type = "circular", same_level = FALSE,
  label_percentage = 1, ...)

\method{plot_dependencies}{default}(x, type = "circular",
  same_level = FALSE, label_percentage = 1, ...)

\method{plot_dependencies}{character}(x, type = "circular",
  same_level = FALSE, label_percentage = 1, ...)

\method{plot_dependencies}{deepdep}(x, type = "circular",
  same_level = FALSE, label_percentage = 1, ...)
}
\arguments{
\item{x}{A \code{\link{deepdep}} object or a \code{character} package name.}

\item{type}{A \code{character}. Possible values are \code{circular} and \code{tree}.}

\item{same_level}{A \code{logical}. If \code{TRUE} links between dependencies on the same
level will be added. By default it's \code{FALSE}.}

\item{label_percentage}{A \code{numeric} value between \code{0} and \code{1}. A fraction
of labels to be displayed. By default it's \code{1} (all labels displayed).}

\item{...}{Other arguments passed to the plotting function.}
}
\value{
A \code{ggplot2, gg, ggraph, deepdep_plot} class object.
}
\description{
Visualize dependency data from a \code{deepdep} object using
\code{ggplot2} and \code{ggraph} packages. Several tree-like layouts are available.
}
\examples{
library(deepdep)

dd <- deepdep("stringr")
plot_dependencies(dd, "tree")

dd2 <- deepdep("cranlogs", depth = 2)
plot_dependencies(dd2, "circular")

plot_dependencies("mice", label_percentage = 0.3, downloads = TRUE)

}
